requirements.txt
pbr>=0.6,!=0.7,<1.0
Babel>=1.3

pecan>=0.4.5
PasteDeploy>=1.5.0
netaddr>=0.7.6
keystonemiddleware>=1.0.0

lxml>=2.3
sqlalchemy>=0.9.4

oslo.config>=1.2.1
oslo.db>=1.0.2
